Add logging to RedisConnector
The code has been updated to integrate a logging module into the RedisConnector class. A logger named 'LOGGER' has been configured to push log events to a StreamHandler and the logging level can be set when initializing a RedisConnector instance. Moreover, an informative message is logged each time a Redis connector is initialized.